chore: 🚚 move files around to use nextjs style
